Make the shortcutsmanager part of the base module
