+++ packages/client/src/default-modules.ts
@@ -49,7 +49,6 @@ import { layoutModule } from './features/layout/layout-module';
 import { navigationModule } from './features/navigation/navigation-module';
 import { routingModule } from './features/routing/routing-module';
 import { selectModule } from './features/select/select-module';
-import { shortcutsModule } from './features/shortcuts/shortcuts-module';
 import { sourceModelWatcherModule } from './features/source-model-watcher/source-model-watcher-module';
 import { statusModule } from './features/status/status-module';
 import { svgMetadataModule } from './features/svg-metadata/svg-metadata-module';
@@ -102,7 +101,6 @@ export const DEFAULT_MODULES = [
     zorderModule,
     svgMetadataModule,
     statusModule,
-    shortcutsModule,
     resizeModule
 ] as const;
