+++ packages/client/src/index.ts
@@ -41,6 +41,10 @@ export * from './base/mouse-position-tracker';
 export * from './base/ranked';
 export * from './base/selection-clearing-mouse-listener';
 export * from './base/selection-service';
+export * from './base/shortcuts/available-shortcuts-extension';
+export * from './base/shortcuts/available-shortcuts-tool';
+export * from './base/shortcuts/shortcuts-manager';
+export * from './base/shortcuts/shortcuts-module';
 export * from './base/tool-manager/tool';
 export * from './base/tool-manager/tool-manager';
 export * from './base/ui-extension/ui-extension';
@@ -166,11 +170,6 @@ export * from './features/save/save-module';
 export * from './features/select/select-feedback-command';
 export * from './features/select/select-module';
 export * from './features/select/select-mouse-listener';
-export * from './features/shortcuts/available-shortcuts-extension';
-export * from './features/shortcuts/available-shortcuts-tool';
-export * from './features/shortcuts/shortcuts-manager';
-export * from './features/shortcuts/shortcuts-model';
-export * from './features/shortcuts/shortcuts-module';
 export * from './features/source-model-watcher/source-model-changed-action-handler';
 export * from './features/source-model-watcher/source-model-watcher-module';
 export * from './features/status/status-module';
